+++ src/host/mod.rs
@@ -6,7 +6,7 @@
 pub mod framing;
 
 cfg_if::cfg_if! {
-    if #[cfg(feature = "heltec_v3_uart")] {
+    if #[cfg(any(feature = "heltec_v3_uart", feature = "elecrow_thinknode_m2"))] {
         mod uart;
         pub use uart::host_task;
     } else {